Refactored Table component to separate business and UI logic for better testability.
